Moved telemetry constants and interface out
